+++ src/roostr.js
@@ -10,10 +10,13 @@
  export { PerspectiveCamera }  from './cameras/perspective';
  export { OrthographicCamera } from './cameras/orthographic';
  export { Mesh }               from './mesh';
+ //export { Entity }             from './entity';
  export { FullscreenQuad }     from './geometries/fullscreenQuad';
+ export { Quad }               from './geometries/quad';
  export { Cube }               from './geometries/cube';
  export { Sphere }             from './geometries/sphere';
  export { VWing }              from './geometries/vwing';
  export { Texture }            from './texture';
+ export { Material }           from './material';
 //export { Matrix4x3 }       from './maths/matrix4x3';
 //export { WebGLRenderer }   from './renderers/WebGLRenderer';